Split PushService into smaller files
